requirements.txt
hikari
hikari-lightbulb
git+https://github.com/MahdiAw/hackernews-python@master
tortoise-orm
